no warnings

+++ Kernel/libs/stack.c
@@ -1,3 +1,6 @@
+// This is a personal academic project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++ and C#: http://www.viva64.com
+
 #include <stack.h>
 
 
@@ -18,11 +21,11 @@ uint64_t load_stack(uint64_t rip, uint64_t rsp, char ** argv, uint64_t argc, uin
     stack * my_stack = (stack *) (rsp - sizeof(stack));
    
     my_stack->regs.rdi = rip;
-    my_stack->regs.rsi = argv;
+    my_stack->regs.rsi = (uint64_t) argv;
     my_stack->regs.rdx = argc;
     my_stack->regs.rcx = pid;
     //my_stack->rip = rip;
-    my_stack->rip = &process_wrapper;
+    my_stack->rip = (uint64_t) &process_wrapper;
     my_stack->cs = 0x8;
     my_stack->rflags = 0x202;
     my_stack->rsp = rsp;
